requirements.txt
requests
python-dotenv
numpy
PyYAML
